requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
bcrypt==4.1.3
PyJWT==2.4.0
Werkzeug==3.0.3